refactor: refactor error middleware

+++ middlewares/errors.ts
@@ -8,7 +8,7 @@ export default (
   next: NextFunction
 ) => {
   if (error) {
-    removeFile(req);
+    req.file && removeFile(req);
     const statusCode = error.status || error.statusCode || 500;
     const message = error.message || "Internal Server Error !!";
     res.status(statusCode).json({ message });
@@ -16,17 +16,15 @@ export default (
 };
 
 function removeFile(req: Request) {
-  if (req.file) {
-    const folders = fs.readdirSync(path.join(process.cwd(), "public"));
+  const folders = fs.readdirSync(path.join(process.cwd(), "public"));
 
-    folders.forEach((folder) => {
-      const files = fs.readdirSync(path.join(process.cwd(), "public", folder));
+  folders.forEach((folder) => {
+    const files = fs.readdirSync(path.join(process.cwd(), "public", folder));
 
-      if (files.includes(req.file!.filename)) {
-        fs.unlinkSync(
-          path.join(process.cwd(), "public", folder, req.file!.filename)
-        );
-      }
-    });
-  }
+    if (files.includes(req.file!.filename)) {
+      fs.unlinkSync(
+        path.join(process.cwd(), "public", folder, req.file!.filename)
+      );
+    }
+  });
 }
